Implementation of PCA(Principal Component Analysis) algorithm
This commit adds an implementation of the PCA (Principal Component Analysis). PCA is a widely used technique for dimensionality reduction in machine learning and data analysis. The implementation is based on the QR algorithm for computing the eigendecomposition of a symmetric matrix. The PCA class takes a matrix of input data and returns a set of eigenvalues and eigenvectors that can be used to transform the data into a lower-dimensional space.